@@ -13,15 +13,17 @@ apply from: 'https://raw.githubusercontent.com/SizableShrimp/Forge-Class-Remappe
 apply plugin: 'eclipse'
 apply plugin: 'idea'
 
-def mc_version = '1.16.4'
-def aklib_version = '1.0.1.0'
-version = "${mc_version}-1.6.5.2" as Object
-group = 'ak.mcmod.chaindestruction' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+def mc_version = '1.17.1'
+def forge_version = '37.1.1'
+def aklib_version = '1.0.2.0'
+def mod_id = 'chaindestruction'
+version = "${mc_version}-1.6.6.0" as Object
+group = 'ak.mcmod' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'chain-destruction'
 tasks.withType(JavaCompile) {
     options.encoding = 'UTF-8'
 }
-java.toolchain.languageVersion = JavaLanguageVersion.of(8)
+java.toolchain.languageVersion = JavaLanguageVersion.of(16)
 
 minecraft {
     // the mappings can be changed at any time, and must be in the following format.
@@ -29,7 +31,7 @@ minecraft {
     // stable_#            stables are built at the discretion of the MCP team.
     // Use non-default mappings at your own risk. they may not always work.
     // simply re-run your setup task after changing the mappings to update your workspace.
-    mappings channel: 'official', version: '1.16.5'
+    mappings channel: 'official', version: mc_version
     // makeObfSourceJar = false // an Srg named sources jar is made by default. uncomment this to disable.
 
     //accessTransformer = file('my_at.cfg')
@@ -75,7 +77,7 @@ minecraft {
             property 'forge.logging.console.level', 'debug'
 
             mods {
-                examplemod {
+                chaindestruction {
                     source sourceSets.main
                 }
             }
@@ -100,7 +102,7 @@ minecraft {
             args '--mod', 'chaindestruction', '--all', '--output', file('src/generated/resources/'), '--existing', file('src/main/resources/')
 
             mods {
-                examplemod {
+                chaindestruction {
                     source sourceSets.main
                 }
             }
@@ -117,7 +119,7 @@ dependencies {
     // Specify the version of Minecraft to use, If this is any group other then 'net.minecraft' it is assumed
     // that the dep is a ForgeGradle 'patcher' dependency. And it's patches will be applied.
     // The userdev artifact is a special name and will get all sorts of transformations applied to it.
-    minecraft 'net.minecraftforge:forge:1.16.4-35.1.37'
+    minecraft "net.minecraftforge:forge:${mc_version}-${forge_version}"
 
     implementation(fg.deobf("ak.mcmod:aklib-${mc_version}:${aklib_version}"))
 }
@@ -128,7 +130,7 @@ jar {
     manifest {
         attributes(["Specification-Title": "chaindestruction",
                     "Specification-Vendor": "ak",
-                    "Specification-Version": "35.0", // We are version 1 of the modlauncher specification
+                    "Specification-Version": "${forge_version}", // We are version 1 of the modlauncher specification
                     "Implementation-Title": project.name,
                     "Implementation-Version": "${archiveVersion}",
                     "Implementation-Vendor" :"ak",
