@@ -2,26 +2,20 @@ package ak.mcmod.chaindestruction.capability;
 
 import ak.mcmod.chaindestruction.ChainDestruction;
 import com.google.common.base.Joiner;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.CapabilityInject;
 import net.minecraftforge.common.capabilities.CapabilityManager;
+import net.minecraftforge.common.capabilities.CapabilityToken;
 
 /**
  * 連鎖破壊ステータスハンドリングクラス
  * Created by A.K. on 2016/09/16.
  */
-public class CapabilityCDPlayerStatusHandler {
+public class CapabilityAdditionalPlayerStatus {
 
   public final static ResourceLocation CD_STATUS = new ResourceLocation(ChainDestruction.MOD_ID, "cd_status");
-  @SuppressWarnings("CanBeFinal")
-  @CapabilityInject(ICDPlayerStatusHandler.class)
-  public static Capability<ICDPlayerStatusHandler> CAPABILITY_CHAIN_DESTRUCTION_PLAYER = null;
+  public static final Capability<IAdditionalPlayerStatus> CAPABILITY = CapabilityManager.get(new CapabilityToken<>() {});
   protected static final Joiner COMMA_JOINER = Joiner.on(',');
-
-  public static void register() {
-    CapabilityManager.INSTANCE.register(ICDPlayerStatusHandler.class, new CDPlayerStatus(), CDPlayerStatus::new);
-  }
 
   /**
    * 連鎖破壊設定をコピーするメソッド
@@ -29,7 +23,7 @@ public class CapabilityCDPlayerStatusHandler {
    * @param copyFrom コピー元
    * @param copyTo   コピー先
    */
-  public static void copyPlayerStatus(ICDPlayerStatusHandler copyFrom, ICDPlayerStatusHandler copyTo) {
+  public static void copyPlayerStatus(IAdditionalPlayerStatus copyFrom, IAdditionalPlayerStatus copyTo) {
     copyTo.setFace(copyFrom.getFace());
     copyTo.setModeType(copyFrom.getModeType());
     copyTo.setDigUnder(copyFrom.isDigUnder());
@@ -46,8 +40,8 @@ public class CapabilityCDPlayerStatusHandler {
    * @param statusHandler 表示させたい設定
    * @return 変換された文字列
    */
-  public static String makePlayerStatusToString(ICDPlayerStatusHandler statusHandler) {
-    StringBuilder sb = new StringBuilder();
+  public static String makePlayerStatusToString(IAdditionalPlayerStatus statusHandler) {
+    var sb = new StringBuilder();
     sb.append("enableItems:[");
     COMMA_JOINER.appendTo(sb, statusHandler.getEnableItems());
     sb.append("]\n");
@@ -56,6 +50,9 @@ public class CapabilityCDPlayerStatusHandler {
     sb.append("]\n");
     sb.append("enableLogBlocks:[");
     COMMA_JOINER.appendTo(sb, statusHandler.getEnableLogBlocks());
+    sb.append("]\n");
+    sb.append("forbiddenTags:[");
+    COMMA_JOINER.appendTo(sb, statusHandler.getForbiddenTags());
     sb.append("]\n");
     return sb.toString();
   }
