@@ -1,16 +1,19 @@
 package ak.mcmod.chaindestruction.capability;
 
+import net.minecraft.nbt.CompoundTag;
+import net.minecraftforge.common.util.INBTSerializable;
+
 import java.util.Set;
 
 /**
  * プライベートモード時のItemStackに保持する情報のインターフェース
  * Created by A.K. on 2016/09/25.
  */
-public interface ICDItemStackStatusHandler {
+public interface IAdditionalItemStackStatus extends INBTSerializable<CompoundTag> {
   /**
    * 連鎖破壊対象ブロックの集合
    *
-   * @return Set<String>
+   * @return Set&lt;String&gt;
    */
   Set<String> getEnableBlocks();
 
@@ -24,7 +27,7 @@ public interface ICDItemStackStatusHandler {
   /**
    * 木こりモード用連鎖破壊対象ブロックの集合
    *
-   * @return Set<String>
+   * @return Set&lt;String&gt;
    */
   Set<String> getEnableLogBlocks();
 
