@@ -1,12 +1,8 @@
 package ak.mcmod.chaindestruction.event;
 
-import ak.mcmod.chaindestruction.command.CommandCopyRtoLCDStatus;
-import ak.mcmod.chaindestruction.command.CommandResetCDPlayerStatus;
-import ak.mcmod.chaindestruction.command.CommandShowItemCDStatus;
-import ak.mcmod.chaindestruction.command.CommandShowPlayerCDStatus;
-import com.mojang.brigadier.CommandDispatcher;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.command.CommandSource;
+import ak.mcmod.chaindestruction.command.CommandItemCDStatus;
+import ak.mcmod.chaindestruction.command.CommandPlayerCDStatus;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
@@ -20,11 +16,9 @@ import javax.annotation.ParametersAreNonnullByDefault;
 //@Mod.EventBusSubscriber(modid = MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE, value = {Dist.DEDICATED_SERVER})
 public class CommandEvents {
   @SubscribeEvent
-  public static void registerCommand(final RegisterCommandsEvent event) {
-    CommandDispatcher<CommandSource> commandDispatcher = event.getDispatcher();
-    CommandCopyRtoLCDStatus.register(commandDispatcher);
-    CommandResetCDPlayerStatus.register(commandDispatcher);
-    CommandShowItemCDStatus.register(commandDispatcher);
-    CommandShowPlayerCDStatus.register(commandDispatcher);
+  public static void registerCommands(final RegisterCommandsEvent event) {
+    var commandDispatcher = event.getDispatcher();
+    CommandItemCDStatus.register(commandDispatcher);
+    CommandPlayerCDStatus.register(commandDispatcher);
   }
 }
