@@ -2,20 +2,19 @@ package ak.mcmod.chaindestruction.network;
 
 import ak.mcmod.ak_lib.util.StringUtils;
 import ak.mcmod.chaindestruction.api.Constants;
-import ak.mcmod.chaindestruction.capability.CDPlayerStatus;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Util;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraftforge.fml.network.NetworkDirection;
-import net.minecraftforge.fml.network.NetworkEvent.Context;
+import ak.mcmod.chaindestruction.capability.CapabilityAdditionalPlayerStatus;
+import net.minecraft.MethodsReturnNonnullByDefault;
+import net.minecraft.Util;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.fmllegacy.network.NetworkDirection;
+import net.minecraftforge.fmllegacy.network.NetworkEvent;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 
@@ -25,7 +24,7 @@ import java.util.function.Supplier;
  */
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class MessageKeyPressedHandler implements BiConsumer<MessageKeyPressed, Supplier<Context>> {
+public class MessageKeyPressedHandler implements BiConsumer<MessageKeyPressed, Supplier<NetworkEvent.Context>> {
 
   /**
    * キーイベント処理。MessageHandlerから呼ばれる
@@ -34,32 +33,32 @@ public class MessageKeyPressedHandler implements BiConsumer<MessageKeyPressed, S
    * @param player プレイヤー
    * @param key    押下キーを表すbyte
    */
-  public static void doKeyEvent(@Nullable ItemStack item, PlayerEntity player, byte key) {
-    CDPlayerStatus.get(player).ifPresent(status -> {
-      String chat;
+  public static void doKeyEvent(@Nullable ItemStack item, Player player, byte key) {
+    player.getCapability(CapabilityAdditionalPlayerStatus.CAPABILITY).ifPresent(status -> {
+      var chat = "";
       if (Objects.isNull(item)) {
         return;
       }
-      if (key == Constants.RegKEY && !item.isEmpty()) {
-        Set<String> enableItems = status.getEnableItems();
-        String uniqueName = StringUtils.getUniqueString(item.getItem().getRegistryName());
+      if (key == Constants.REG_KEY && !item.isEmpty()) {
+        var enableItems = status.getEnableItems();
+        var uniqueName = StringUtils.getUniqueString(item.getItem().getRegistryName());
         if (player.isShiftKeyDown() && enableItems.contains(uniqueName)) {
           enableItems.remove(uniqueName);
           chat = String.format("Remove Tool : %s", uniqueName);
-          player.sendMessage(new StringTextComponent(chat), Util.NIL_UUID);
+          player.sendMessage(new TextComponent(chat), Util.NIL_UUID);
         }
         if (!player.isShiftKeyDown() && !enableItems.contains(uniqueName)) {
           enableItems.add(uniqueName);
           chat = String.format("Add Tool : %s", uniqueName);
-          player.sendMessage(new StringTextComponent(chat), Util.NIL_UUID);
+          player.sendMessage(new TextComponent(chat), Util.NIL_UUID);
         }
       }
-      if (key == Constants.DigKEY) {
+      if (key == Constants.DIG_KEY) {
         status.setDigUnder(!status.isDigUnder());
         chat = String.format("Dig Under %b", status.isDigUnder());
-        player.sendMessage(new StringTextComponent(chat), Util.NIL_UUID);
+        player.sendMessage(new TextComponent(chat), Util.NIL_UUID);
       }
-      if (key == Constants.ModeKEY) {
+      if (key == Constants.MODE_KEY) {
         if (player.isShiftKeyDown()) {
           status.setPrivateRegisterMode(!status.isPrivateRegisterMode());
           chat = String.format("Private Register Mode %b", status.isPrivateRegisterMode());
@@ -67,17 +66,17 @@ public class MessageKeyPressedHandler implements BiConsumer<MessageKeyPressed, S
           status.setModeType(status.getModeType().getNextModeType());
           chat = String.format("Mode %s", status.getModeType().name());
         }
-        player.sendMessage(new StringTextComponent(chat), Util.NIL_UUID);
+        player.sendMessage(new TextComponent(chat), Util.NIL_UUID);
       }
-      PacketHandler.INSTANCE.sendTo(new MessageCDStatusProperties(player),
-              ((ServerPlayerEntity) player).connection.getConnection(),
+      PacketHandler.INSTANCE.sendTo(new MessageSyncAdditionalPayerStatus(player),
+              ((ServerPlayer) player).connection.getConnection(),
               NetworkDirection.PLAY_TO_CLIENT);
     });
   }
 
   @Override
-  public void accept(MessageKeyPressed messageKeyPressed, Supplier<Context> contextSupplier) {
-    PlayerEntity player = contextSupplier.get().getSender();
+  public void accept(MessageKeyPressed messageKeyPressed, Supplier<NetworkEvent.Context> contextSupplier) {
+    Player player = contextSupplier.get().getSender();
     if (Objects.nonNull(player) && !player.getMainHandItem().isEmpty()) {
       doKeyEvent(player.getMainHandItem(), player, messageKeyPressed.getKey());
     }
