@@ -2,8 +2,8 @@ package ak.mcmod.chaindestruction.util;
 
 import ak.mcmod.chaindestruction.ChainDestruction;
 import com.google.common.collect.Lists;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.MethodsReturnNonnullByDefault;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.ForgeConfigSpec.BooleanValue;
 import net.minecraftforge.common.ForgeConfigSpec.Builder;
@@ -11,7 +11,7 @@ import net.minecraftforge.common.ForgeConfigSpec.ConfigValue;
 import net.minecraftforge.common.ForgeConfigSpec.IntValue;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.config.ModConfig.Loading;
+import net.minecraftforge.fml.event.config.ModConfigEvent;
 import org.apache.logging.log4j.LogManager;
 
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -32,13 +32,13 @@ public class ConfigUtils {
   public static final ForgeConfigSpec CONFIG_SPEC;
 
   static {
-    Builder builder = new ForgeConfigSpec.Builder();
+    var builder = new ForgeConfigSpec.Builder();
     COMMON = new Common(builder);
     CONFIG_SPEC = builder.build();
   }
 
   @SubscribeEvent
-  public static void configLoading(final Loading event) {
+  public static void configLoading(final ModConfigEvent.Loading event) {
     LogManager.getLogger().debug("Loaded ChainDestruction config file {}",
             event.getConfig().getFileName());
     COMMON.maxYforTreeMode = COMMON.maxYforTreeModeConfigValue.get();
@@ -59,6 +59,11 @@ public class ConfigUtils {
 
   public static class Common {
 
+    private final IntValue maxYforTreeModeConfigValue;
+    private final BooleanValue destroyingSequentiallyConfigValue;
+    private final IntValue digTaskMaxCounterConfigValue;
+    private final BooleanValue notToDestroyItemConfigValue;
+    private final ConfigValue<String> excludeRegisterItemsConfigValue;
     public int maxYforTreeMode;
     public boolean destroyingSequentially;
     public int digTaskMaxCounter;
@@ -67,11 +72,6 @@ public class ConfigUtils {
     public boolean dropOnPlayer = true;
     public List<String> excludeRegisterItemList;
     public Predicate<ResourceLocation> excludeItemPredicate;
-    private final IntValue maxYforTreeModeConfigValue;
-    private final BooleanValue destroyingSequentiallyConfigValue;
-    private final IntValue digTaskMaxCounterConfigValue;
-    private final BooleanValue notToDestroyItemConfigValue;
-    private final ConfigValue<String> excludeRegisterItemsConfigValue;
 
     Common(Builder builder) {
       builder.comment("Common settings")
